Resolve undefined and refactor filenames
