+++ ormconfig.json
@@ -7,7 +7,7 @@
   "database": "virtuose",
   "synchronize": true,
   "logging": false,
-  "entities": ["src/entities/**/*.ts"],
+  "entities": ["src/entities/**/*.ts", "src/modules/**/*.entity.ts"],
   "migrations": ["src/migrations/**/*.ts"],
   "subscribers": ["src/subscribers/**/*.ts"],
   "cli": {
