+++ package.json
@@ -7,6 +7,7 @@
   "author": "Sylvain Colignon <sylvain.colignon@gmail.com>",
   "license": "Apache-2.0",
   "devDependencies": {
+    "@types/bcrypt": "^3.0.0",
     "@types/node": "^10.11.4",
     "nodemon": "^1.18.4",
     "ts-node": "^7.0.1",
@@ -14,6 +15,7 @@
     "typescript": "3.1.2"
   },
   "dependencies": {
+    "bcrypt": "^3.0.1",
     "graphql-yoga": "^1.16.2",
     "mysql": "^2.16.0",
     "pg": "^7.5.0",
