+++ src/index.ts
@@ -1,4 +1,9 @@
 import { App } from './app';
 
 const app = new App();
-app.start();
+app
+  .start()
+  .then(() => {})
+  .catch(error => {
+    console.log(error);
+  });
